requirements.txt
opencv-python
numpy
matplotlib
pytesseract
Pillow